+++ src/roam-navigator.js
@@ -37,6 +37,10 @@
   // Key to toggle left sidebar visibility.
   const LEFT_SIDEBAR_KEY = '`';
 
+  // Maximum number of breadcrumbs (recent pages) to keep track of /
+  // attempt to display.
+  const MAX_BREADCRUMB_COUNT = 15;
+
   function readSetting(name, initial) {
     if (window.roamNavigatorSettings && name in window.roamNavigatorSettings) {
       return roamNavigatorSettings[name];
@@ -56,6 +60,15 @@
   // Set to true to respond to scroll keys outside navigate mode.
   const SCROLL_OUTSIDE_NAVIGATE_MODE =
         readSetting('scroll-outside-navigate-mode', true);
+
+  // Set to true to enable display of recent pages list.
+  const BREADCRUMBS_ENABLED =
+        readSetting('breadcrumbs-enabled', true);
+
+  // Set to true to display recent pages list, even when not in
+  // navigation mode.
+  const BREADCRUMBS_ALWAYS_VISIBLE =
+        readSetting('breadcrumbs-always-visible', false);
 
   // 'navigate' (g) attempts to assign keys to items based on their
   // names. In some case there might not be a concise labeling. This
@@ -169,6 +182,7 @@
         handleFocusOut();
       }
       blockWasHighlighted = blockHighlighted;
+      updateBreadcrumbs();
     });
 
     const observer = new MutationObserver(() => {
@@ -307,10 +321,10 @@
     currentNavigatePrefixesUsed = {};
     navigateKeysPressed = '';
 
-
-    // TODO: Cleanup: Was a function to disconnect observer, but now
-    // it is always registered.
     finishNavigate = () => {
+      if (!BREADCRUMBS_ALWAYS_VISIBLE) {
+        clearBreadcrumbs();
+      }
     };
 
     setupNavigate(false);
@@ -337,6 +351,7 @@
   // overrides the keyboard handler such that it temporarily expects a
   // key.
   function setupNavigate(onlyLinks) {
+    updateBreadcrumbs();
     // ensureSidebarOpen();
     if (!matchingClass(NAVIGATE_CLASS)(document.body)) {
       document.body.classList.add(NAVIGATE_CLASS);
@@ -382,7 +397,7 @@
     }
   }
 
-  function collectNavigateOptions(onlyLinks) {
+  function collectNavigateOptions() {
     const sidebar = getUniqueClass(document, 'roam-sidebar-container');
     // Initialize a list of elements to bind to keys for
     // navigation. Starts out with some reserved keys that will
@@ -557,6 +572,11 @@
       addLinks(linksByUid, navigateOptions, rightSidebar);
     });
 
+    const breadcrumbsContainer = getUniqueClass(document, BREADCRUMBS_CLASS);
+    if (breadcrumbsContainer) {
+      addLinks(linksByUid, navigateOptions, breadcrumbsContainer);
+    }
+
     const linkItems = [];
     for (const uid of Object.keys(linksByUid)) {
       linkItems.push(linksByUid[uid]);
@@ -569,16 +589,20 @@
   }
 
 
-  function addLinks(linksByUid, navigateOptions, el) {
-    const links = el.querySelectorAll([
+  function addLinks(linksByUid, navigateOptions, container) {
+    const links = container.querySelectorAll([
       '.rm-page-ref',
       'a',
     ].join(', '));
     for (let i = 0; i < links.length; i++) {
       const link = links[i];
       const boundingRect = link.getBoundingClientRect();
-      if (boundingRect.bottom > 50 &&
-          boundingRect.top < window.innerHeight - 10) {
+      const visible =
+            container.classList.contains(BREADCRUMBS_CLASS) ?
+            boundingRect.top < 45 :
+            boundingRect.bottom > 50 &&
+              boundingRect.top < window.innerHeight - 10;
+      if (visible) {
         const parent = link.parentElement;
         let el;
         let uid;
@@ -652,6 +676,7 @@
   function rerenderTips(onlyLinks) {
     let renderedAny = false;
     withDomMutation(() => {
+      updateBreadcrumbs();
       // ensureSidebarOpen();
       removeOldTips(onlyLinks);
       if (!onlyLinks) {
@@ -688,14 +713,14 @@
   }
 
   function renderTipInternal(prefix, rest, el, extraClasses) {
-    const tip = div(HINT_CLASS, text(rest));
+    const tip = div({'class': HINT_CLASS}, text(rest));
     if (extraClasses) {
       for (const cls of extraClasses) {
         tip.classList.add(cls);
       }
     }
     if (prefix.length > 0) {
-      tip.prepend(span(HINT_TYPED_CLASS, text(prefix)));
+      tip.prepend(span({'class': HINT_TYPED_CLASS}, text(prefix)));
     }
     if (matchingClass('rm-block-text')(el) ||
         el.id === 'block-input-ghost') {
@@ -1216,6 +1241,146 @@
     return finishNavigate !== null;
   }
 
+  /*****************************************************************************
+   * Recent history breadcrumbs
+   */
+
+  // MUTABLE. Array of recently visited pages. { title: "", hash: "", uid: "" }
+  const breadcrumbs = [];
+
+  // Class used for breadcrumbs container.
+  const BREADCRUMBS_CLASS = 'roam_navigator_breadcrumbs';
+
+  // Attributes used for link portion of breadcrumbs.
+  const LINK_ATTRS = {
+    'tabindex': '-1',
+    'class': 'rm-page-ref rm-page-ref-link-color',
+  };
+
+  // Regex used to extract id from hash portion of url.
+  const ID_FROM_HASH_REGEX = /#\/app\/[^\/]*\/page\/([a-zA-Z0-9\-]*)$/;
+
+  // Regex used to identify if the hash portion of the url is the daily page.
+  const IS_DAILY_NOTES_REGEX = /#\/app\/[^\/]*$/;
+
+  function updateBreadcrumbs() {
+    if (BREADCRUMBS_ENABLED) {
+      const hash = window.location.hash;
+      const pageTitleElement =
+            document.querySelector('.roam-body-main .rm-title-display > span');
+      const pageUidMatchResult = ID_FROM_HASH_REGEX.exec(hash);
+      const isDailyPage = IS_DAILY_NOTES_REGEX.exec(hash) !== null;
+      let title;
+      let uid;
+      if (pageTitleElement) {
+        title = pageTitleElement.innerText;
+      } else if (!isDailyPage) {
+        if (!pageUidMatchResult) {
+          // Fall back on using document title, this is used for cases
+          // like the graph overview / all pages.
+          title = document.title;
+          uid = document.title;
+        } else {
+          debug('Didn\'t find title element for page');
+          return;
+        }
+      }
+      if (isDailyPage) {
+        uid = 'daily_notes';
+        title = 'Daily Notes';
+      } else {
+        if (pageUidMatchResult) {
+          uid = pageUidMatchResult[1];
+        }
+      }
+      let newBreadcrumb;
+      if (uid) {
+        newBreadcrumb = {hash, title, uid};
+      } else {
+        newBreadcrumb = {hash, title};
+      }
+      let changed = false;
+      if (breadcrumbs.length < 1) {
+        breadcrumbs.push(newBreadcrumb);
+        changed = true;
+      }
+      const lastBreadcrumb = breadcrumbs[breadcrumbs.length - 1];
+      if (lastBreadcrumb.hash === hash) {
+        if (lastBreadcrumb.title !== title) {
+          lastBreadcrumb.title = title;
+        }
+      } else {
+        // Remove existing occurrences.
+        for (let i = 0; i < breadcrumbs.length; i++) {
+          if (breadcrumbs[i].hash === hash) {
+            breadcrumbs.splice(i, 1);
+          }
+        }
+        breadcrumbs.push(newBreadcrumb);
+        changed = true;
+      }
+      if (changed) {
+        trimExcessBreadcrumbs();
+        debug('updated breadcrumbs = ', breadcrumbs);
+      }
+      // Update rendering of breadcrumbs.
+      const alreadyVisible =
+            selectAll(document, '.' + BREADCRUMBS_CLASS).length > 0;
+      const shouldBeVisible = BREADCRUMBS_ALWAYS_VISIBLE || isNavigating();
+      if (!shouldBeVisible) {
+        clearBreadcrumbs();
+      } else if (changed || !alreadyVisible) {
+        clearBreadcrumbs();
+        renderBreadcrumbs();
+      }
+    }
+  }
+
+  function clearBreadcrumbs() {
+    withDomMutation(() => {
+      withQuery(document, '.' + BREADCRUMBS_CLASS, (container) => {
+        container.parentNode.removeChild(container);
+      });
+    });
+  }
+
+  function renderBreadcrumbs() {
+    withDomMutation(() => {
+      const container = div({'class': BREADCRUMBS_CLASS});
+      for (let i = breadcrumbs.length - 2; i >= 0; i--) {
+        const breadcrumb = breadcrumbs[i];
+        const breadcrumbAttrs = {
+          'title': breadcrumb.title,
+          'data-link-title': breadcrumb.title,
+        };
+        if ('uid' in breadcrumb) {
+          breadcrumbAttrs['data-link-uid'] = breadcrumb['uid'];
+        }
+        const breadcrumbSpan = span(breadcrumbAttrs);
+        breadcrumbSpan.appendChild(span(LINK_ATTRS, text(breadcrumb.title)));
+        breadcrumbSpan.onclick = (event) => {
+          window.location.hash = breadcrumb.hash;
+        };
+        container.appendChild(breadcrumbSpan);
+      }
+
+      withUnique(document, '.roam-topbar > .flex-h-box', (topbar) => {
+        const buttonClasses = ['bp3-icon-menu', 'bp3-icon-menu-open'];
+        const sidebarButton = getUniqueClass(topbar, buttonClasses);
+        if (sidebarButton && sidebarButton.nextSibling) {
+          topbar.insertBefore(container, sidebarButton.nextSibling);
+        } else {
+          topbar.insertBefore(container, topbar.firstChild);
+        }
+      });
+    });
+  }
+
+  function trimExcessBreadcrumbs() {
+    if (breadcrumbs.length > MAX_BREADCRUMB_COUNT) {
+      breadcrumbs.splice(0, breadcrumbs.length - MAX_BREADCRUMB_COUNT);
+    }
+  }
 
   function withDomMutation(f) {
     domMutationLevel += 1;
@@ -1413,10 +1578,9 @@
   }
 
   // Uses querySelectorAll, and applies the provided function to each result.
-  // eslint-disable-next-line no-unused-vars
   function withQuery(parent, query, f) {
     const els = selectAll(parent, query);
-    for (const i = 0; i < els.length; i++) {
+    for (let i = 0; i < els.length; i++) {
       f(els[i]);
     }
   }
@@ -1656,10 +1820,10 @@
     return element('div', ...rest);
   }
 
-  function element(t, cls, ...children) {
+  function element(t, attrs, ...children) {
     const el = document.createElement(t);
-    if (cls) {
-      el.classList.add(cls);
+    for (const attr of Object.keys(attrs)) {
+      el.setAttribute(attr, attrs[attr]);
     }
     for (const child of children) {
       el.appendChild(child);
@@ -1712,9 +1876,32 @@
     '  display: inline;',
     '  margin-top: -14px;',
     '}',
+    '.' + BREADCRUMBS_CLASS + ' .' + LINK_HINT_CLASS + ' {',
+    '  margin-top: -2px;',
+    '  margin-left: 5px;',
+    '}',
     // Prevents clipping of tips.
     '.' + NAVIGATE_CLASS + ' .parent-path-wrapper {',
+    '  flex: 100 0 0;',
     '  overflow: visible !important;',
+    '}',
+    '.' + BREADCRUMBS_CLASS + ' {',
+    '  flex: 100 0 0;',
+    '  overflow: hidden;',
+    '  height: 45px;',
+    '  line-height: 45px;',
+    '}',
+    '.' + BREADCRUMBS_CLASS + ' > span {',
+    '  float: right;',
+    '  margin-left: 5px;',
+    '}',
+    '.' + BREADCRUMBS_CLASS + ' .rm-page-ref {',
+    '  white-space: nowrap;',
+    '  overflow: hidden;',
+    '  max-width: 256px;',
+    '  text-overflow: ellipsis;',
+    '  border-left: 0.5px solid #666;',
+    '  padding-left: 5px;',
     '}',
   ].join('\n'));